backend/DAtabase_architecture/requirements.txt
supabase
pymongo
transformers
torch
PyMuPDF
langgraph
google-generativeai
